Early implementation of package versioning.
The primary package version is stored in VERSION.  It should be updated after every release.  For now, setup.py will generate a unique version number for every build by appending a date/timestamp to the version number from VERSION.  That's enough to allow each build to be pushed to testpypi.  A more elegant version numbering system will have to be implemented before releasing to pypi proper.